+++ pystream/pipeline/serial_pipeline.py
@@ -12,10 +12,23 @@ class SerialPipeline(PipelineBase):
     def __init__(
         self,
         stages: List[StageCallable],
+        names: List[Optional[str]],
         profiler_handler: Optional[ProfilerHandler] = None,
     ) -> None:
+        """The class that will handle the serial pipeline.
+
+        Args:
+            stages (List[StageCallable]): The stages to be run
+                in sequence.
+            names (List[Optional[str]]): Stage names. If the name is None,
+                default stage name will be given.
+            profiler_handler (Optional[ProfilerHandler]): Handler for the profiler.
+                If None, no profiling attempt will be done.
+        """
         self.final_stage = FinalStage(profiler_handler)
-        self.pipeline: List[Stage] = [StageContainer(stage) for stage in stages]
+        self.pipeline: List[Stage] = [
+            StageContainer(stage, name) for stage, name in zip(stages, names)
+        ]
         self.pipeline.append(self.final_stage)
         self.results = PipelineData()